ex-010/requirements.txt
pyyaml
scikit-learn
scipy
matplotlib
textblob
networkx
pystache
